requirements.txt
streamlit
requests
azure-maps-search
azure-core
opencv-python
numpy
Pillow
pandas
